Print string

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -6,7 +6,8 @@
  */
 void _puts(char *str)
 {
-	int i;
+	long int i;
+
 
 	for (i = 0; i < strlen(str); i++)
 		_putchar(str[i]);
